+++ crates/fm/src/file_reader.rs
@@ -7,7 +7,8 @@ use std::path::Path;
 
 #[derive(RustEmbed)]
 #[folder = "../../noir_stdlib/src"]
-#[prefix = "std/"]
+#[cfg_attr(not(target_os = "windows"), prefix = "std/")]
+#[cfg_attr(target_os = "windows", prefix = r"std\")] // Note reversed slash direction
 struct StdLibAssets;
 
 cfg_if::cfg_if! {
@@ -25,9 +26,7 @@ cfg_if::cfg_if! {
         pub(crate) fn read_file_to_string(path_to_file: &Path) -> Result<String, Error> {
             use std::io::ErrorKind;
 
-            let path_str = path_to_file.as_os_str().to_str().unwrap();
-
-            match StdLibAssets::get(path_str) {
+            match StdLibAssets::get(path_to_file.to_str().unwrap()) {
 
                 Some(std_lib_asset) => {
                     Ok(std::str::from_utf8(std_lib_asset.data.as_ref()).unwrap().to_string())